Implement user login, registration, and contact ownership

+++ controllers/contactController.js
@@ -2,7 +2,7 @@ import mongoose from "mongoose";
 import Contact from "../models/contactModel.js";
 
 export async function getContacts(req, res) {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id:req.user.id});
   return res.status(200).json(contacts);
 }
 
@@ -13,7 +13,7 @@ export async function createContact(req, res) {
     throw new Error("All fields are mandatory");
   }
 
-  const contact = await Contact.create({ name, email, phoneNumber });
+  const contact = await Contact.create({ name, email, phoneNumber, user_id: req.user.id });
 
   return res.status(201).json(contact);
 }
@@ -40,6 +40,10 @@ export async function updateContact(req, res) {
   if (!contact) {
     return res.status(404).json({ error: "Contact ID not found" });
   }
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403)
+    throw new Error("User not authorised for this operation")
+  }
 
   const updatedContact = await Contact.findByIdAndUpdate(id, req.body, {
     new: true,
@@ -56,6 +60,10 @@ export async function deleteContact(req, res) {
   const contact = await Contact.findById(id);
   if (!contact) {
     return res.status(404).json({ error: "Contact ID not found" });
+  }
+   if(contact.user_id.toString() !== req.user.id){
+    res.status(403)
+    throw new Error("User not authorised for this operation")
   }
 
   await Contact.findByIdAndDelete(id);
